# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "回忆小卖部 AI 角色明信片交流应用 API 文档",
        "title": "Memory Postcard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/postcards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取用户的明信片列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "明信片"
                ],
                "summary": "获取明信片列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "对话ID",
                        "name": "conversation_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "character_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "draft",
                            "sent",
                            "delivered",
                            "read"
                        ],
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否收藏",
                        "name": "is_favorite",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "default": "created_at",
                        "description": "排序字段",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "排序方向",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.PaginatedResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "发送明信片给AI角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "明信片"
                ],
                "summary": "创建明信片",
                "parameters": [
                    {
                        "description": "明信片信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PostcardCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Postcard"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/postcards/conversations/{conversation_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据对话ID获取完整的对话记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "明信片"
                ],
                "summary": "获取对话记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "对话ID",
                        "name": "conversation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Postcard"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/postcards/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取明信片的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "明信片"
                ],
                "summary": "获取明信片详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "明信片ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Postcard"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新明信片信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "明信片"
                ],
                "summary": "更新明信片",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "明信片ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PostcardUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Postcard"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除明信片",
                "tags": [
                    "明信片"
                ],
                "summary": "删除明信片",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "明信片ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Character": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "creator": {
                    "description": "关联关系",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "creator_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "drafts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Draft"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "popularity_score": {
                    "type": "number"
                },
                "postcards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Postcard"
                    }
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserCharacterRelation"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "usage_count": {
                    "type": "integer"
                },
                "user_role_desc": {
                    "type": "string"
                },
                "user_role_name": {
                    "description": "用户角色设定",
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                },
                "voices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CharacterVoice"
                    }
                }
            }
        },
...
        "models.CharacterVoice": {
            "type": "object",
            "properties": {
                "character": {
                    "$ref": "#/definitions/models.Character"
                },
                "character_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "file_format": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "voice_url": {
                    "type": "string"
                }
            }
        },
        "models.Draft": {
            "type": "object",
            "properties": {
                "character": {
                    "$ref": "#/definitions/models.Character"
                },
                "character_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "emotion_tags": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "landscape_image_url": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Favorite": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "favoritable_id": {
                    "type": "integer"
                },
                "favoritable_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "models.Postcard": {
            "type": "object",
            "properties": {
                "ai_generated_image_url": {
                    "type": "string"
                },
                "character": {
                    "$ref": "#/definitions/models.Character"
                },
                "character_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "postcard_template": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "关联关系",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.PostcardCreateRequest": {
            "type": "object",
            "required": [
                "character_id",
                "content"
            ],
            "properties": {
                "character_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string",
                    "maxLength": 36
                },
                "image_url": {
                    "type": "string",
                    "maxLength": 255
                },
                "postcard_template": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "models.PostcardUpdateRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string",
                    "maxLength": 255
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "postcard_template": {
                    "type": "string",
                    "maxLength": 100
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "sent",
                        "delivered",
                        "read"
                    ]
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "characters": {
                    "description": "关联关系",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Character"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "dark_mode": {
                    "type": "boolean"
                },
                "drafts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Draft"
                    }
                },
                "email": {
                    "type": "string"
                },
                "favorites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Favorite"
                    }
                },
                "font_size": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "postcards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Postcard"
                    }
                },
                "signature": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserCharacterRelation": {
            "type": "object",
            "properties": {
                "character": {
                    "$ref": "#/definitions/models.Character"
                },
                "character_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interaction_count": {
                    "type": "integer"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "last_interaction_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}